hotel_finale_project/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
